cs_code/Demo.LearnByDoing-5afb7d3fae028348316b4ede0c8d969d9c1fe84c/Demo.LearnByDoing.Tests/Chapter01/Chapter1_8.cs
using System;
using System.Collections.Generic;

namespace Demo.LearnByDoing.Tests.Chapter01
{
    public class Chapter1_8
    {
        public IEnumerable<Tuple<int, int>> GetZeroPositions(int[,] matrix)
        {
            for (int i = 0; i < matrix.GetLength(0); i++)
            {
                for (int j = 0; j < matrix.GetLength(1); j++)
                {
                    if (matrix[i, j] == 0)
                        yield return new Tuple<int, int>(i, j);
                }
            }
        }
    }
}